+++ test/test_load_case.jl
@@ -1,17 +1,19 @@
 import HVDCWISE_TEA as _HWTEA
 
 const _HWTEA_dir = dirname(dirname(pathof(_HWTEA))) # Root directory of HVDCWISE_TEA package
+const save_results = true  # To save test results in Excel files
 
 using Test
 include("../src/io/load_case.jl")
+
 @testset "Tests Cases" begin
-    #
+    
     @testset "Test case 01 : Single AC Line" begin
         # Load and run test case 1
-        global results_1 = load_case("test01")
+        baseMVA = 100
+        global results_1 = load_case("test01", baseMVA, true)
 
         if results_1 !== nothing
-            baseMVA = 100
             @test results_1["termination_status"] == _HWTEA.OPTIMAL
             @test results_1["objective"] ≈ 9900.0 rtol = 1e-3
             # All results are in per unit, with baseMVA = 100 MVA (specified in the .m file)
@@ -62,38 +64,37 @@ include("../src/io/load_case.jl")
         end
     end
     =#
-    @testset "Test case 05a : 1 storage + 1 generator + 1 load" begin
-        # Load and run test case 5a
-        global results_5 = load_case("test05a")
-
-        ## baseMVA reading
-        file = joinpath(_HWTEA_dir, "test\\data\\test05a\\.m_files\\test05a.m")
-        file_string_data = read(open(file), String)
-        file_dict_data = _IM.parse_matlab_string(file_string_data, extended=true)[1];
-        baseMVA = file_dict_data["mpc.baseMVA"]  # 100 MVA
+    
+    @testset "Test case 05 : 1 storage + 1 generator + 1 load" begin
+        # Load and run test case 5
+        baseMVA = 100
+        global results_5 = load_case("test05", baseMVA, true)
 
         if results_5 !== nothing
             @test results_5["termination_status"] == _HWTEA.OPTIMAL
             @test results_5["objective"] ≈ 10450 rtol = 1e-3
             # t=1: Demand=8, Available=10 and storage is available -> 1 MW consumed by the storage & 8 MW by the load (pgcurt=1 MW)
             @test results_5["solution"]["nw"]["1"]["gen"]["1"]["pg"] ≈ 9/baseMVA atol=1e-3
-            @test results_5["solution"]["nw"]["1"]["gen"]["1"]["pgcurt"] ≈ 1/baseMVA atol=1e-3
+            # @test results_5["solution"]["nw"]["1"]["gen"]["1"]["pgcurt"] ≈ 1/baseMVA atol=1e-3  # no pgcurt because dispatchable generator
             @test results_5["solution"]["nw"]["1"]["load"]["1"]["pcurt"] ≈ 0/baseMVA atol=1e-3
-            @test results_5["solution"]["nw"]["1"]["storage"]["1"]["ps"] ≈ -1/baseMVA atol=1e-3  # FIXME what is the storage attribute for storage net production ?
+            @test results_5["solution"]["nw"]["1"]["storage"]["1"]["ps"] ≈ -1/baseMVA atol=1e-3  # ps=sc-sd (storage power = chargin power - discharging power)
+            @test results_5["solution"]["nw"]["1"]["storage"]["1"]["sd"] ≈ 1/baseMVA atol=1e-3  # ps=sc-sd
 
             # t=2: Demand=12, Available=10 but storage not available -> 2 MW of load shedding
             @test results_5["solution"]["nw"]["2"]["gen"]["1"]["pg"] ≈ 10/baseMVA atol=1e-3
-            @test results_5["solution"]["nw"]["2"]["gen"]["1"]["pgcurt"] ≈ 0/baseMVA atol=1e-3
+            # @test results_5["solution"]["nw"]["2"]["gen"]["1"]["pgcurt"] ≈ 0/baseMVA atol=1e-3
             @test results_5["solution"]["nw"]["2"]["load"]["1"]["pcurt"] ≈ 2/baseMVA atol=1e-3
-            @test results_5["solution"]["nw"]["2"]["storage"]["1"]["ps"] ≈ 0/baseMVA atol=1e-3
+            @test !haskey(results_5["solution"]["nw"]["2"], "storage")  # no storage in the results because status=0
 
             # t=3: Demand=12, Available=10 and storage is available -> 1 MW produced by the storage & 1 MW of load shedding
             @test results_5["solution"]["nw"]["3"]["gen"]["1"]["pg"] ≈ 10/baseMVA atol=1e-3
-            @test results_5["solution"]["nw"]["3"]["gen"]["1"]["pgcurt"] ≈ 0/baseMVA atol=1e-3
+            # @test results_5["solution"]["nw"]["3"]["gen"]["1"]["pgcurt"] ≈ 0/baseMVA atol=1e-3
             @test results_5["solution"]["nw"]["3"]["load"]["1"]["pcurt"] ≈ 1/baseMVA atol=1e-3
             @test results_5["solution"]["nw"]["3"]["storage"]["1"]["ps"] ≈ 1/baseMVA atol=1e-3
+            @test results_5["solution"]["nw"]["1"]["storage"]["1"]["sc"] ≈ 1/baseMVA atol=1e-3  # ps=sc-sd
         else
-            println("Error loading test05a. Check CSV and .m files name and location.")
+            println("Error loading test05. Check CSV and .m files name and location.")
         end
     end
+
 end